fix: remove billing v1
