+++ frontend/src/lib/components/BillingAlerts.tsx
@@ -1,17 +1,17 @@
 import { useActions, useValues } from 'kea'
 import { router } from 'kea-router'
 import { useEffect, useState } from 'react'
-import { billingV2Logic } from 'scenes/billing/v2/billingV2Logic'
+import { billingLogic } from 'scenes/billing/billingLogic'
 import { urls } from 'scenes/urls'
 import { AlertMessage } from 'lib/lemon-ui/AlertMessage'
 
-export function BillingAlertsV2(): JSX.Element | null {
-    const { billingAlert, billingVersion } = useValues(billingV2Logic)
-    const { reportBillingAlertShown } = useActions(billingV2Logic)
+export function BillingAlerts(): JSX.Element | null {
+    const { billingAlert, billingVersion } = useValues(billingLogic)
+    const { reportBillingAlertShown } = useActions(billingLogic)
     const { currentLocation } = useValues(router)
     const [alertHidden, setAlertHidden] = useState(false)
 
-    const showAlert = billingAlert && billingVersion === 'v2'
+    const showAlert = billingAlert && billingVersion === ''
 
     useEffect(() => {
         if (showAlert) {
@@ -19,7 +19,7 @@ export function BillingAlertsV2(): JSX.Element | null {
         }
     }, [showAlert])
 
-    if (!billingAlert || billingVersion !== 'v2' || alertHidden) {
+    if (!billingAlert || billingVersion !== '' || alertHidden) {
         return null
     }
