+++ frontend/src/scenes/appScenes.ts
@@ -38,7 +38,6 @@ export const appScenes: Record<Scene, () => any> = {
     [Scene.ProjectCreateFirst]: () => import('./project/Create'),
     [Scene.SystemStatus]: () => import('./instance/SystemStatus'),
     [Scene.ToolbarLaunch]: () => import('./toolbar-launch/ToolbarLaunch'),
-    [Scene.Licenses]: () => import('./instance/Licenses'),
     [Scene.AsyncMigrations]: () => import('./instance/AsyncMigrations/AsyncMigrations'),
     [Scene.DeadLetterQueue]: () => import('./instance/DeadLetterQueue/DeadLetterQueue'),
     [Scene.MySettings]: () => import('./me/Settings'),
@@ -48,7 +47,6 @@ export const appScenes: Record<Scene, () => any> = {
     [Scene.InviteSignup]: () => import('./authentication/InviteSignup'),
     [Scene.Ingestion]: () => import('./ingestion/IngestionWizard'),
     [Scene.Billing]: () => import('./billing/Billing'),
-    [Scene.BillingSubscribed]: () => import('./billing/BillingSubscribed'),
     [Scene.Plugins]: () => import('./plugins/Plugins'),
     [Scene.FrontendAppScene]: () => import('./apps/FrontendAppScene'),
     [Scene.AppMetrics]: () => import('./apps/AppMetricsScene'),
