--- frontend/src/scenes/billing/Billing.scss
@@ -1,29 +1,0 @@
-.BillingPlan {
-    max-width: 500px;
-    flex-grow: 1;
-    .BillingPlan__description {
-        ol,
-        ul {
-            margin-top: 1rem;
-            padding-left: 0;
-            list-style: none;
-            text-align: center;
-            li {
-                line-height: 1.2rem;
-                margin-bottom: 1rem;
-            }
-            .disclaimer {
-                font-size: 10px;
-                font-weight: bold;
-                vertical-align: top;
-                padding-left: 1px;
-            }
-        }
-
-        .disclaimer-details {
-            font-size: 0.75rem;
-            color: var(--muted);
-            margin-top: 1rem;
-        }
-    }
-}
