+++ frontend/src/scenes/billing/PlanTable.scss
