+++ frontend/src/scenes/billing/billingLogic.ts
@@ -1,236 +1,256 @@
-import { kea, path, actions, connect, reducers, selectors, events, listeners } from 'kea'
+import { kea, path, actions, connect, reducers, afterMount, selectors, listeners } from 'kea'
 import { loaders } from 'kea-loaders'
 import api from 'lib/api'
-import type { billingLogicType } from './billingLogicType'
-import { PlanInterface, BillingType } from '~/types'
-import { preflightLogic } from 'scenes/PreflightCheck/preflightLogic'
-import posthog from 'posthog-js'
-import { sceneLogic } from 'scenes/sceneLogic'
-import { Scene } from 'scenes/sceneTypes'
-import { lemonToast } from 'lib/lemon-ui/lemonToast'
+import { BillingProductType, BillingType } from '~/types'
+import { router } from 'kea-router'
 import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
-import { windowValues } from 'kea-window-values'
-import { getBreakpoint } from 'lib/utils/responsiveUtils'
 import { urlToAction } from 'kea-router'
-import { urls } from 'scenes/urls'
-import { eventUsageLogic } from 'lib/utils/eventUsageLogic'
-import { billingV2Logic } from './v2/billingV2Logic'
+import { forms } from 'kea-forms'
+import { dayjs } from 'lib/dayjs'
+import { lemonToast } from '@posthog/lemon-ui'
+import { projectUsage } from './billing-utils'
+import posthog from 'posthog-js'
+import { preflightLogic } from 'scenes/PreflightCheck/preflightLogic'
+import { userLogic } from 'scenes/userLogic'
+import { pluralize } from 'lib/utils'
+import type { billingLogicType } from './billingLogicType'
 
-export const UTM_TAGS = 'utm_medium=in-product&utm_campaign=billing-management'
 export const ALLOCATION_THRESHOLD_ALERT = 0.85 // Threshold to show warning of event usage near limit
+export const ALLOCATION_THRESHOLD_BLOCK = 1.2 // Threshold to block usage
 
-export enum BillingAlertType {
-    SetupBilling = 'setup_billing',
-    UsageNearLimit = 'usage_near_limit',
-    UsageLimitExceeded = 'usage_limit_exceeded',
+export interface BillingAlertConfig {
+    status: 'info' | 'warning' | 'error'
+    title: string
+    message: string
+    contactSupport?: boolean
+}
+
+const parseBillingResponse = (data: Partial<BillingType>): BillingType => {
+    if (data.billing_period) {
+        data.billing_period = {
+            current_period_start: dayjs(data.billing_period.current_period_start),
+            current_period_end: dayjs(data.billing_period.current_period_end),
+            interval: data.billing_period.interval,
+        }
+
+        data.products?.forEach((x) => {
+            x.projected_usage = projectUsage(x.current_usage, data.billing_period)
+        })
+    }
+
+    data.free_trial_until = data.free_trial_until ? dayjs(data.free_trial_until) : undefined
+
+    return data as BillingType
 }
 
 export const billingLogic = kea<billingLogicType>([
     path(['scenes', 'billing', 'billingLogic']),
     actions({
+        setShowLicenseDirectInput: (show: boolean) => ({ show }),
+        reportBillingAlertShown: (alertConfig: BillingAlertConfig) => ({ alertConfig }),
+        reportBillingShown: true,
         registerInstrumentationProps: true,
-        toggleUsageTiers: true,
-        setPlans: (plans: PlanInterface[]) => ({ plans }),
-        referer: (referer: string) => ({ referer }),
     }),
     connect({
-        values: [preflightLogic, ['preflight'], featureFlagLogic, ['featureFlags'], billingV2Logic, ['billingVersion']],
-        actions: [eventUsageLogic, ['reportIngestionBillingCancelled']],
+        values: [featureFlagLogic, ['featureFlags'], preflightLogic, ['preflight']],
+        actions: [userLogic, ['loadUser']],
     }),
     reducers({
-        showUsageTiers: [
-            false as boolean,
+        showLicenseDirectInput: [
+            false,
             {
-                toggleUsageTiers: (state) => !state,
-            },
-        ],
-        billingSuccessRedirect: [
-            urls.projectHomepage() as string,
-            {
-                referer: (_, { referer }) => {
-                    if (referer === 'ingestion') {
-                        return urls.events()
-                    } else {
-                        return urls.projectHomepage()
-                    }
-                },
+                setShowLicenseDirectInput: (_, { show }) => show,
             },
         ],
     }),
-    windowValues({
-        isSmallScreen: (window: Window) => window.innerWidth < getBreakpoint('md'),
-    }),
-    loaders(({ actions }) => ({
+    loaders(() => ({
         billing: [
             null as BillingType | null,
             {
                 loadBilling: async () => {
-                    const response = await api.get('api/billing/')
-                    if (!response?.plan) {
-                        actions.loadPlans()
-                    } else {
-                        actions.setPlans([response.plan])
-                    }
-                    actions.registerInstrumentationProps()
-                    return response as BillingType
-                },
-                setBillingLimit: async (billing: BillingType) => {
-                    const res = await api.update('api/billing/', billing)
-                    lemonToast.success(`Billing limit set to $${billing.billing_limit} usd/month`)
+                    const response = await api.get('api/billing-v2')
 
-                    return res as BillingType
+                    return parseBillingResponse(response)
+                },
+
+                updateBillingLimits: async (limits: { [key: string]: string | null }) => {
+                    const response = await api.update('api/billing-v2', { custom_limits_usd: limits })
+
+                    lemonToast.success('Billing limits updated')
+                    return parseBillingResponse(response)
                 },
             },
         ],
-        plans: [
-            [] as PlanInterface[],
+        products: [
+            [] as BillingProductType[],
             {
-                loadPlans: async () => {
-                    const response = await api.get('api/plans?self_serve=1')
-                    return response.results
-                },
-                setPlans: ({ plans }) => plans,
-            },
-        ],
-        billingSubscription: [
-            null as BillingType | null,
-            {
-                subscribe: async (plan) => {
-                    return await api.create('billing/subscribe', { plan })
-                },
-            },
-        ],
-        planDetails: [
-            null as string | null,
-            {
-                loadPlanDetails: async (plan) => {
-                    const response = await fetch(`/api/plans/${plan}/template/`)
-                    if (response.ok) {
-                        return await response.text()
-                    }
-                    return null
+                loadProducts: async () => {
+                    const response = await api.get('api/billing-v2/available_products')
+                    return response
                 },
             },
         ],
     })),
     selectors({
-        upgradeLink: [
-            (s) => [s.preflight, s.billingVersion],
-            (preflight, billingVersion): string =>
-                billingVersion === 'v2' || preflight?.cloud
-                    ? '/organization/billing'
-                    : 'https://license.posthog.com?utm_medium=in-product&utm_campaign=in-product-upgrade',
-        ],
-        eventAllocation: [(s) => [s.billing], (billing: BillingType) => billing?.event_allocation],
-        percentage: [
-            (s) => [s.eventAllocation, s.billing],
-            (eventAllocation: number | null, billing: BillingType) => {
-                if (!eventAllocation || !billing?.current_usage) {
-                    return null
+        billingAlert: [
+            (s) => [s.billing, s.preflight],
+            (billing, preflight): BillingAlertConfig | undefined => {
+                if (!billing || !preflight?.cloud) {
+                    return
                 }
-                return Math.min(Math.round((billing.current_usage / eventAllocation) * 100) / 100, 1)
-            },
-        ],
-        strokeColor: [
-            (selectors) => [selectors.percentage],
-            (percentage) => {
-                let color: string | Record<string, string> = 'var(--primary)'
-                if (percentage === null || percentage === undefined) {
-                    /* No event limit set */
-                    color = {
-                        from: '#1890FF',
-                        to: '#52C41A',
+
+                if (billing.free_trial_until && billing.free_trial_until.isAfter(dayjs())) {
+                    const remainingDays = billing.free_trial_until.diff(dayjs(), 'days')
+                    const remainingHours = billing.free_trial_until.diff(dayjs(), 'hours')
+
+                    if (remainingHours > 72) {
+                        return
+                    }
+
+                    return {
+                        status: 'info',
+                        title: `Your free trial will end in ${
+                            remainingHours < 24 ? pluralize(remainingHours, 'hour') : pluralize(remainingDays, 'day')
+                        }.`,
+                        message: `Setup billing now to ensure you don't lose access to premium features.`,
                     }
                 }
 
-                if (percentage && percentage > 0.65 && percentage < 0.8) {
-                    color = 'var(--warning)'
-                }
-                if (percentage && percentage > 0.8) {
-                    color = 'var(--danger)'
-                }
-                return color
-            },
-        ],
-        alertToShow: [
-            (s) => [s.eventAllocation, s.percentage, s.billing, sceneLogic.selectors.scene, s.billingVersion],
-            (
-                eventAllocation: number | null,
-                percentage: number,
-                billing: BillingType,
-                scene: Scene,
-                billingVersion: string
-            ): BillingAlertType | undefined => {
-                if (billingVersion === 'v2') {
-                    return
-                }
-                if (billing?.should_setup_billing && billing?.subscription_url) {
-                    // Determines which billing alert/warning to show to the user (if any)
-
-                    // Priority 1: In-progress incomplete billing setup
-                    return BillingAlertType.SetupBilling
+                if (billing.deactivated) {
+                    return {
+                        status: 'error',
+                        title: 'Your organization has been temporarily suspended.',
+                        message: 'Please contact support to reactivate it.',
+                        contactSupport: true,
+                    }
                 }
 
-                // Priority 2: Event allowance exceeded or near limit
-                if (billing?.billing_limit_exceeded) {
-                    return BillingAlertType.UsageLimitExceeded
+                const productOverLimit = billing.products?.find((x) => {
+                    return x.percentage_usage > 1
+                })
+
+                if (productOverLimit) {
+                    return {
+                        status: 'error',
+                        title: 'Usage limit exceeded',
+                        message: `You have exceeded the usage limit for ${productOverLimit.name}. Please upgrade your plan or data loss may occur.`,
+                    }
                 }
 
-                // Priority 3: Event allowance near threshold
-                if (
-                    scene !== Scene.Billing &&
-                    billing?.is_billing_active &&
-                    billing?.current_usage &&
-                    eventAllocation &&
-                    percentage >= ALLOCATION_THRESHOLD_ALERT
-                ) {
-                    return BillingAlertType.UsageNearLimit
+                const productApproachingLimit = billing.products?.find(
+                    (x) => x.percentage_usage > ALLOCATION_THRESHOLD_ALERT
+                )
+
+                if (productApproachingLimit) {
+                    return {
+                        status: 'info',
+                        title: 'You will soon hit your usage limit',
+                        message: `You have currently used ${parseFloat(
+                            (productApproachingLimit.percentage_usage * 100).toFixed(2)
+                        )}% of your ${productApproachingLimit.type.toLowerCase()} allocation.`,
+                    }
                 }
             },
         ],
     }),
-    events(({ actions }) => ({
-        afterMount: () => {
-            if (preflightLogic.values.preflight?.cloud) {
-                actions.loadBilling()
-            }
+    forms(({ actions }) => ({
+        activateLicense: {
+            defaults: { license: '' } as { license: string },
+            errors: ({ license }) => ({
+                license: !license ? 'Please enter your license key' : undefined,
+            }),
+            submit: async ({ license }, breakpoint) => {
+                breakpoint(500)
+                try {
+                    await api.update('api/billing-v2/license', {
+                        license,
+                    })
+
+                    // Reset the URL so we don't trigger the license submission again
+                    router.actions.replace('/organization/billing?success=true')
+                    setTimeout(() => {
+                        window.location.reload() // Permissions, projects etc will be out of date at this point, so refresh
+                    }, 100)
+                } catch (e: any) {
+                    actions.setActivateLicenseManualErrors({
+                        license: e.detail || 'License could not be activated. Please contact support.',
+                    })
+                    throw e
+                }
+            },
         },
     })),
-    listeners(({ values }) => ({
-        subscribeSuccess: ({ billingSubscription }) => {
-            if (billingSubscription?.subscription_url) {
-                window.location.href = billingSubscription.subscription_url
+
+    listeners(({ actions, values }) => ({
+        reportBillingShown: () => {
+            posthog.capture('billing v2 shown')
+        },
+        reportBillingAlertShown: ({ alertConfig }) => {
+            posthog.capture('billing alert shown', {
+                ...alertConfig,
+            })
+        },
+        loadBillingSuccess: () => {
+            if (
+                router.values.location.pathname.includes('/organization/billing') &&
+                router.values.searchParams['success']
+            ) {
+                // if the activation is successful, we reload the user to get the updated billing info on the organization
+                actions.loadUser()
+                router.actions.replace('/organization/billing')
             }
+            actions.registerInstrumentationProps()
         },
         registerInstrumentationProps: async (_, breakpoint) => {
             await breakpoint(100)
             if (posthog && values.billing) {
-                posthog.register({
-                    has_billing_plan: !!values.billing?.plan,
-                    metered_billing: values.billing.plan?.is_metered_billing,
-                    event_allocation: values.billing.event_allocation,
-                    allocation_used:
-                        values.billing.event_allocation && values.billing.current_usage !== null
-                            ? values.billing.current_usage / values.billing.event_allocation
-                            : undefined,
-                })
-            }
-        },
-
-        loadBillingSuccess: () => {
-            if (!values.billing) {
-                return
+                const payload = {
+                    has_billing_plan: !!values.billing.has_active_subscription,
+                    free_trial_until: values.billing.free_trial_until?.toISOString(),
+                    customer_deactivated: values.billing.deactivated,
+                    current_total_amount_usd: values.billing.current_total_amount_usd,
+                }
+                if (values.billing.custom_limits_usd) {
+                    for (const product of Object.keys(values.billing.custom_limits_usd)) {
+                        payload[`custom_limits_usd.${product}`] = values.billing.custom_limits_usd[product]
+                    }
+                }
+                if (values.billing.products) {
+                    for (const product of values.billing.products) {
+                        const type = product.type.toLowerCase()
+                        payload[`percentage_usage.${type}`] = product.percentage_usage
+                        payload[`current_amount_usd.${type}`] = product.current_amount_usd
+                        payload[`unit_amount_usd.${type}`] = product.unit_amount_usd
+                        payload[`usage_limit.${type}`] = product.usage_limit
+                        payload[`current_usage.${type}`] = product.current_usage
+                        payload[`projected_usage.${type}`] = product.projected_usage
+                        payload[`free_allocation.${type}`] = product.free_allocation
+                    }
+                }
+                if (values.billing.billing_period) {
+                    payload['billing_period_start'] = values.billing.billing_period.current_period_start
+                    payload['billing_period_end'] = values.billing.billing_period.current_period_end
+                }
+                if (values.billing.license) {
+                    payload['license_plan'] = values.billing.license.plan
+                }
+                posthog.register(payload)
             }
         },
     })),
+
+    afterMount(({ actions }) => {
+        actions.loadBilling()
+    }),
+
     urlToAction(({ actions }) => ({
-        '/ingestion/billing': (_, params) => {
-            if (params.reason === 'cancelled') {
-                actions.reportIngestionBillingCancelled()
+        // IMPORTANT: This needs to be above the "*" so it takes precedence
+        '/organization/billing': (_params, _search, hash) => {
+            if (hash.license) {
+                actions.setShowLicenseDirectInput(true)
+                actions.setActivateLicenseValues({ license: hash.license })
+                actions.submitActivateLicense()
             }
-        },
-        '/organization/billing/subscribed': (_, { referer }) => {
-            actions.referer(referer)
         },
     })),
 ])
