+++ frontend/src/scenes/scenes.ts
@@ -212,9 +212,6 @@ export const sceneConfigurations: Partial<Record<Scene, SceneConfig>> = {
         instanceLevel: true,
         name: 'Instance status & settings',
     },
-    [Scene.Licenses]: {
-        instanceLevel: true,
-    },
     [Scene.AsyncMigrations]: {
         instanceLevel: true,
     },
@@ -229,10 +226,6 @@ export const sceneConfigurations: Partial<Record<Scene, SceneConfig>> = {
     [Scene.Billing]: {
         hideProjectNotice: true,
         organizationBased: true,
-    },
-    [Scene.BillingSubscribed]: {
-        plain: true,
-        allowUnauthenticated: true,
     },
     [Scene.Unsubscribe]: {
         allowUnauthenticated: true,
@@ -335,10 +328,8 @@ export const routes: Record<string, Scene> = {
     [urls.projectCreateFirst()]: Scene.ProjectCreateFirst,
     [urls.organizationSettings()]: Scene.OrganizationSettings,
     [urls.organizationBilling()]: Scene.Billing,
-    [urls.billingSubscribed()]: Scene.BillingSubscribed,
     [urls.organizationCreateFirst()]: Scene.OrganizationCreateFirst,
     [urls.organizationCreationConfirm()]: Scene.OrganizationCreationConfirm,
-    [urls.instanceLicenses()]: Scene.Licenses,
     [urls.instanceStatus()]: Scene.SystemStatus,
     [urls.instanceSettings()]: Scene.SystemStatus,
     [urls.instanceStaffUsers()]: Scene.SystemStatus,
