+++ frontend/src/scenes/session-recordings/player/sessionRecordingPlayerLogic.ts
@@ -29,7 +29,7 @@ import { lemonToast } from '@posthog/lemon-ui'
 import { delay } from 'kea-test-utils'
 import { ExportedSessionRecordingFile } from '../file-playback/sessionRecordingFilePlaybackLogic'
 import { userLogic } from 'scenes/userLogic'
-import { openBillingPopupModal } from 'scenes/billing/v2/BillingPopup'
+import { openBillingPopupModal } from 'scenes/billing/BillingPopup'
 import { sessionRecordingsListLogic } from 'scenes/session-recordings/playlist/sessionRecordingsListLogic'
 import { router } from 'kea-router'
 import { urls } from 'scenes/urls'
