+++ frontend/src/scenes/session-recordings/playlist/playlistUtils.ts
@@ -9,7 +9,7 @@ import { lemonToast } from 'lib/lemon-ui/lemonToast'
 import { DEFAULT_RECORDING_FILTERS } from 'scenes/session-recordings/playlist/sessionRecordingsListLogic'
 import { router } from 'kea-router'
 import { urls } from 'scenes/urls'
-import { openBillingPopupModal } from 'scenes/billing/v2/BillingPopup'
+import { openBillingPopupModal } from 'scenes/billing/BillingPopup'
 import { PLAYLIST_LIMIT_REACHED_MESSAGE } from 'scenes/session-recordings/sessionRecordingsLogic'
 
 function getOperatorSymbol(operator: PropertyOperator | null): string {
